feat(skip-dialog): Enhance popup usability and design (#118)
changed workflow for skipping cookies
snackbar appears when skip_cookies is set
remove icon
update messages
