+++ Phonebook.Frontend/src/app/pages/page-information/page-information.component.html
@@ -53,7 +53,6 @@
           This site uses cookies, to enhance your experience. Cookies are only used for necessary tasks, e.g. saving the
           language you want to use.
         </span>
-        <span i18n="Page-informationComponent|NoCockieUrl@@PageInformationComponentSectionNoCookies"></span>
       </mat-card-content>
     </mat-card>
   </div>
