+++ Phonebook.Frontend/src/i18n/messages.en.xlf
@@ -993,11 +993,19 @@
         <note priority="1" from="meaning">Display-notificationDialog</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesText" datatype="html">
         <source>
-        Annoyed by the start-dialogs? Klick here:
-      </source><target state="final">
-        Annoyed by the start-dialogs? Klick here:
-      </target>
+      You deactivated cookies and keep getting those dialogs?
+    </source><target state="final">
+      You deactivated cookies and keep getting those dialogs?
+    </target>
         <note priority="1" from="description">Annoyed by cookies notice</note>
+        <note priority="1" from="meaning">Display-notificationDialog</note>
+      </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesButton" datatype="html">
+        <source>
+      Deactivate startup dialogues!
+    </source><target state="final">
+      Deactivate startup dialogues!
+    </target>
+        <note priority="1" from="description">Deactivate Startup Dialogues</note>
         <note priority="1" from="meaning">Display-notificationDialog</note>
       </trans-unit><trans-unit id="PageInformationComponentSectionCookies" datatype="html">
         <source>
@@ -1322,13 +1330,13 @@
         <note priority="1" from="description">Title of Table Column "Building"</note>
         <note priority="1" from="meaning">TableComponent</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookies" datatype="html">
-        <source>Add skip_dialog=true as parameter to url for skipping start dialogs if cookies are deleted after each session.</source><target state="final">Add skip_dialog=true as parameter to url for skipping start dialogs if cookies are deleted after each session.</target>
+        <source>Please notice: You wont get any information about updates or releases with the set url prameter.</source><target state="final">Please notice: You wont get any information about updates or releases with the set url prameter.</target>
         <note priority="1" from="description">Message for new no cookie url</note>
-        <note priority="1" from="meaning">Display-New-Url-Advice</note>
+        <note priority="1" from="meaning">Display advice to new url</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesUrl" datatype="html">
-        <source>Go to Url</source><target state="final">Go to Url</target>
+        <source>Reset Url</source><target state="final">Reset Url</target>
         <note priority="1" from="description">Message for following no cookie url</note>
-        <note priority="1" from="meaning">Go-to-New-Url-Advice</note>
+        <note priority="1" from="meaning">Restore Url</note>
       </trans-unit><trans-unit id="DataPersonStatus" datatype="html">
         <source>
               Status
