+++ Phonebook.Frontend/src/i18n/messages.test.xlf
@@ -877,9 +877,15 @@
         <note priority="1" from="meaning">Display-notificationDialog</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesText" datatype="html">
         <source>
-        Annoyed by the start-dialogs? Klick here:
-      </source><target state="final">TEST</target>
+      You deactivated cookies and keep getting those dialogs?
+    </source><target state="translated">TEST</target>
         <note priority="1" from="description">Annoyed by cookies notice</note>
+        <note priority="1" from="meaning">Display-notificationDialog</note>
+      </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesButton" datatype="html">
+        <source>
+      Deactivate startup dialogues!
+    </source><target state="translated">TEST</target>
+        <note priority="1" from="description">Deactivate Startup Dialogues</note>
         <note priority="1" from="meaning">Display-notificationDialog</note>
       </trans-unit><trans-unit id="PageInformationComponentSectionCookies" datatype="html">
         <source>
@@ -1105,13 +1111,13 @@
         <note priority="1" from="description">Title of Table Column "Building"</note>
         <note priority="1" from="meaning">TableComponent</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookies" datatype="html">
-        <source>Add skip_dialog=true as parameter to url for skipping start dialogs if cookies are deleted after each session.</source><target state="translated">TEST</target>
+        <source>Please notice: You wont get any information about updates or releases with the set url prameter.</source><target state="translated">TEST</target>
         <note priority="1" from="description">Message for new no cookie url</note>
-        <note priority="1" from="meaning">Display-New-Url-Advice</note>
+        <note priority="1" from="meaning">Display advice to new url</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesUrl" datatype="html">
-        <source>Go to Url</source><target state="final">TEST</target>
+        <source>Reset Url</source><target state="translated">TEST</target>
         <note priority="1" from="description">Message for following no cookie url</note>
-        <note priority="1" from="meaning">Go-to-New-Url-Advice</note>
+        <note priority="1" from="meaning">Restore Url</note>
       </trans-unit><trans-unit id="DataPersonStatus" datatype="html">
         <source>
               Status
