+++ Phonebook.Frontend/src/i18n/messages.xlf
@@ -104,9 +104,15 @@
         <note priority="1" from="meaning">Display-notificationDialog</note>
       </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesText" datatype="html">
         <source>
-        Annoyed by the start-dialogs? Klick here:
-      </source>
+      You deactivated cookies and keep getting those dialogs?
+    </source>
         <note priority="1" from="description">Annoyed by cookies notice</note>
+        <note priority="1" from="meaning">Display-notificationDialog</note>
+      </trans-unit><trans-unit id="PageInformationNewUrlNoCookiesButton" datatype="html">
+        <source>
+      Deactivate startup dialogues!
+    </source>
+        <note priority="1" from="description">Deactivate Startup Dialogues</note>
         <note priority="1" from="meaning">Display-notificationDialog</note>
       </trans-unit><trans-unit id="BugReportConsentComponentTitle" datatype="html">
         <source>
@@ -900,6 +906,16 @@
         <note priority="1" from="description">Attributions SubTitle</note>
         <note priority="1" from="meaning">Page-informationComponent</note>
       </trans-unit>
+      <trans-unit id="PageInformationNewUrlNoCookies" datatype="html">
+        <source>Please notice: You wont get any information about updates or releases with the set url prameter.</source>
+        <note priority="1" from="description">Message for new no cookie url</note>
+        <note priority="1" from="meaning">Display advice to new url</note>
+      </trans-unit>
+      <trans-unit id="PageInformationNewUrlNoCookiesUrl" datatype="html">
+        <source>Reset Url</source>
+        <note priority="1" from="description">Message for following no cookie url</note>
+        <note priority="1" from="meaning">Restore Url</note>
+      </trans-unit>
       <trans-unit id="OrganigramNodeComponentCopiedFirstPart" datatype="html">
         <source>Link to</source>
         <note priority="1" from="description">First part of the message displayed when copying a link to the node</note>
@@ -1089,16 +1105,6 @@
         <source>Building</source>
         <note priority="1" from="description">Title of Table Column &quot;Building&quot;</note>
         <note priority="1" from="meaning">TableComponent</note>
-      </trans-unit>
-      <trans-unit id="PageInformationNewUrlNoCookies" datatype="html">
-        <source>Add skip_dialog=true as parameter to url for skipping start dialogs if cookies are deleted after each session.</source>
-        <note priority="1" from="description">Message for new no cookie url</note>
-        <note priority="1" from="meaning">Display-New-Url-Advice</note>
-      </trans-unit>
-      <trans-unit id="PageInformationNewUrlNoCookiesUrl" datatype="html">
-        <source>Go to Url</source>
-        <note priority="1" from="description">Message for following no cookie url</note>
-        <note priority="1" from="meaning">Go-to-New-Url-Advice</note>
       </trans-unit>
       <trans-unit id="GeneralSuccessMessageCopy" datatype="html">
         <source>Copied to clipboard!</source>
